requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-ollama>=0.3.0
langgraph>=0.2.0
fastapi>=0.115.0
openai>=1.0.0
python-dotenv>=1.0.0
prompt-toolkit>=3.0.0
pydantic>=2.0.0
redis>=5.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
typing-extensions>=4.0.0
uvicorn>=0.30.0
